Upgrade qulice to 0.35.2 and fix all warnings
Qulice 0.35.2 brings several new checks that this codebase did not
satisfy, 194 violations in total. All of them are fixed here, with no
suppressions added.

JavadocEmptyLineBeforeTagCheck (140) wants an empty Javadoc line between
a description and the block of at-clauses that follows it. One is
inserted everywhere it was missing.

NoJavadocForPrivateMethodsCheck (47) prohibits Javadoc above a private
method. The mechanical blocks, the ones that only restated the signature
in prose, are gone. Where a block carried an explanation that the name
and the body do not give away, that explanation stays as a plain block
comment above the method, without the at-clauses.

ImplicitConstructorCheck (3) asks for an explicit constructor in a
public class that Javadoc documents, so the three JUnit extension
classes now declare the constructor JUnit calls reflectively.

SingleUseConstantCheck (2) reports a private constant that is mentioned
only once. STAGES and RACES are inlined into their only usage in
RandomPipeline, and what they documented moves into the comment above
walk(), since a comment inside a method body is prohibited.

ProhibitStaticNestedClassesCheck (1) asks for a static nested class in
its own file, so Summary.Collector becomes Collector, taking the private
same() helper with it, as that was its only user.

FullyQualifiedTypeCheck (1) reports a redundant fully qualified name, so
AbstractMojo imports Locale instead of spelling it out.
