+++ pom.xml
@@ -404,7 +404,7 @@
           <plugin>
             <groupId>com.qulice</groupId>
             <artifactId>qulice-maven-plugin</artifactId>
-            <version>0.30.13</version>
+            <version>0.35.2</version>
             <configuration>
               <excludes>
                 <!--
