+++ src/main/java/org/eolang/hone/Summary.java
@@ -4,15 +4,10 @@
  */
 package org.eolang.hone;
 
-import com.jcabi.log.Logger;
 import java.io.IOException;
-import java.nio.file.FileSystemLoopException;
 import java.nio.file.FileVisitOption;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
@@ -20,6 +15,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Build summary statistics.
+ *
  * @since 0.1.0
  */
 public final class Summary {
@@ -36,6 +32,7 @@ public final class Summary {
 
     /**
      * Constructor.
+     *
      * @param root Root directory to search for statistics
      */
     Summary(final Path root) {
@@ -44,6 +41,7 @@ public final class Summary {
 
     /**
      * Constructor.
+     *
      * @param root Root directory to search for statistics
      * @param target Directory to save the summary report
      */
@@ -54,6 +52,7 @@ public final class Summary {
 
     /**
      * Collects summary statistics from all child modules.
+     *
      * @return The path to the generated summary report
      */
     Path collect() {
@@ -65,7 +64,7 @@ public final class Summary {
                 this.root,
                 EnumSet.of(FileVisitOption.FOLLOW_LINKS),
                 Integer.MAX_VALUE,
-                new Summary.Collector(found, stats, destination)
+                new Collector(found, stats, destination)
             );
         } catch (final IOException exception) {
             throw new IllegalStateException(
@@ -90,79 +89,5 @@ public final class Summary {
             "ID",
             ignore -> String.format("%d/%d", index.getAndIncrement(), res.size())
         );
-    }
-
-    /**
-     * Whether two paths point to the same file.
-     * @param first First path
-     * @param second Second path
-     * @return TRUE when both resolve to the same normalized absolute path
-     */
-    private static boolean same(final Path first, final Path second) {
-        return first.toAbsolutePath().normalize()
-            .equals(second.toAbsolutePath().normalize());
-    }
-
-    /**
-     * A file visitor that gathers statistics files, following symlinks.
-     * @since 0.1.0
-     */
-    private static final class Collector extends SimpleFileVisitor<Path> {
-
-        /**
-         * Where to put the found CSVs.
-         */
-        private final List<CSV> found;
-
-        /**
-         * The statistics file name to look for.
-         */
-        private final String stats;
-
-        /**
-         * The summary output, excluded from the walk.
-         */
-        private final Path output;
-
-        /**
-         * Ctor.
-         * @param found Where to put the found CSVs
-         * @param stats The statistics file name to look for
-         * @param output The summary output to exclude
-         */
-        Collector(
-            final List<CSV> found,
-            final String stats,
-            final Path output
-        ) {
-            this.found = found;
-            this.stats = stats;
-            this.output = output;
-        }
-
-        @Override
-        public FileVisitResult visitFile(
-            final Path file, final BasicFileAttributes attrs
-        ) {
-            if (this.stats.equals(file.getFileName().toString())
-                && !Summary.same(this.output, file)) {
-                this.found.add(new CSV(file));
-            }
-            return FileVisitResult.CONTINUE;
-        }
-
-        @Override
-        public FileVisitResult visitFileFailed(
-            final Path file, final IOException exc
-        ) {
-            if (exc instanceof FileSystemLoopException) {
-                Logger.warn(
-                    this,
-                    "Symlink cycle detected at %s, skipping",
-                    file
-                );
-            }
-            return FileVisitResult.CONTINUE;
-        }
     }
 }
