+++ src/main/java/org/eolang/hone/package-info.java
@@ -4,6 +4,7 @@
  */
 /**
  * The main classes of the hone-maven-plugin.
+ *
  * @since 0.1.0
  */
 package org.eolang.hone;
