+++ src/test/java/org/eolang/hone/DisabledWithoutDocker.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 /**
  * This annotation disables JUnit tests when Docker is
  * not runnable in command line.
+ *
  * @since 0.1.0
  */
 @Target({ ElementType.TYPE, ElementType.METHOD })
