+++ src/test/java/org/eolang/hone/DisabledWithoutPhino.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 /**
  * This annotation disables JUnit tests when phino is
  * not runnable in command line.
+ *
  * @since 0.1.0
  */
 @Target({ ElementType.TYPE, ElementType.METHOD })
