+++ src/test/java/org/eolang/hone/MktempTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 
 /**
  * Test case for {@link Mktemp}.
+ *
  * @since 0.1.0
  */
 final class MktempTest {
