+++ src/test/java/org/eolang/hone/RmiMojoTest.java
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
  * Test case for {@link RmiMojo}.
+ *
  * @since 0.1.0
  */
 @ExtendWith(MktmpResolver.class)
