+++ src/test/java/org/eolang/hone/SummaryTest.java
@@ -18,6 +18,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
  * Test case for {@link Rules}.
+ *
  * @since 0.1.0
  */
 @ExtendWith(MktmpResolver.class)
