+++ src/test/java/org/eolang/hone/TimingsTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 
 /**
  * Test case for {@link Timings}.
+ *
  * @since 0.1.0
  */
 final class TimingsTest {
